backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-generativeai>=0.8.0
python-dotenv==1.0.0
protobuf>=5.0.0
reportlab==4.0.9
